Isolate search behavior to its own class
In order to allow not allocating the search button when search is disabled,
I need to isolate the behavior to its own class.
